Work on class file class
The actual class for a class file is mostly kind of done. I can now read in a
basic class file and store the data in a ClassFile class (I think this is what
the big boys call serialization). It's pretty hideous right now, but oh well.
Still learning the Ruby Way™.